backend/python/requirements.txt
boto3==1.34.19
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
reportlab==4.0.9
